refactor: 코드 리팩토링 - index.js : 이벤트 리스너들을 하나의 함수로 모으고 앱 구동 시 실행하도록 수정 - renamed) 'LottoApp.js' to 'createLotto.js' : 로또를 생성한다는 의미에 어울리도록 파일명 변경
