+++ src/js/createLotto.js
